Rewrite BitSet serializer to make if more efficient
The new serializer is not serializing an entire byte per bit.
To access the internal state of BitSet, I have to use Java reflection. The Java 1.7 implementation of BitSet allows to access the internal long[], but chill is compatible to Java 1.6, so we have to play some tricks here.

My experiments show that the new code is 4 times more space efficient (514 vs 2052 bytes) and 11 times faster (446 vs 4908 ms) [both in 100000 ser-deser rounds]