Chill-Akka sample configuration (#304)
Configuring Akka to use Kryo for all `scala.Product` classes may be problematic because many of our internal messages are also case classes, and are expected to be serialized with our internal serializer (e.g. `InitJoin` with `ClusterMessageSerializer`).

Since our marker interfaces do not inherit `scala.Product` but do inherit `java.io.Serializable`, the latter is safer to reconfigure (because of the 'prefer the most specific serializer' semantics). The additional benefit is that this unlocks disabling Java serialization entirely.